refactor: chessGameDao 인터페이스와 구현체 형태로 변경
