+++ src/main/java/chess/controller/command/MoveCommand.java
@@ -1,6 +1,7 @@
 package chess.controller.command;
 
 import chess.dao.ChessGameDao;
+import chess.dao.ChessGameDaoImpl;
 import chess.domain.game.ChessGame;
 import chess.domain.piece.property.Color;
 import chess.domain.position.Position;
@@ -17,7 +18,7 @@ public final class MoveCommand implements Command {
     private static final int TARGET_INDEX = 1;
 
     private final OutputView outputView = new OutputView();
-    private final ChessGameDao chessGameDao = ChessGameDao.getInstace();
+    private final ChessGameDao chessGameDaoImpl = ChessGameDaoImpl.getInstance();
 
     private final List<String> parameters;
 
@@ -47,11 +48,11 @@ public final class MoveCommand implements Command {
         Position source = PositionParser.parse(parameters.get(SOURCE_INDEX));
         Position target = PositionParser.parse(parameters.get(TARGET_INDEX));
         chessGame.playTurn(source, target);
-        chessGameDao.updateGame(chessGame, source, target);
+        chessGameDaoImpl.updateGame(chessGame, source, target);
         outputView.printBoard(chessGame.getBoard());
         if (chessGame.computeWinner() != Color.NONE) {
             outputView.printWinner(chessGame.computeWinner());
-            chessGameDao.deleteGameById(chessGame.getGameId());
+            chessGameDaoImpl.deleteGameById(chessGame.getGameId());
             chessGame.end();
         }
     }
